Major Upgrade to CiviCRM 4.7
* Major Upgrade! Upgrading CiviCRM from 4.6.x to 4.7.x. Be careful! Backup your database, read the upgrade steps in README.md, and fully test on a dev environment before upgrading. Don't merge this code yet if you've got other updates which are urgent. If you have questions contact http://civicrmstarterkit.org/contact. We provide some basic general support for the public. If you require help with your specific website there will likely be a cost.

+++ profiles/civicrm_starterkit/modules/civicrm/CRM/Contact/Page/DedupeRules.php
@@ -1,9 +1,9 @@
 <?php
 /*
  +--------------------------------------------------------------------+
- | CiviCRM version 4.6                                                |
+ | CiviCRM version 4.7                                                |
  +--------------------------------------------------------------------+
- | Copyright CiviCRM LLC (c) 2004-2015                                |
+ | Copyright CiviCRM LLC (c) 2004-2017                                |
  +--------------------------------------------------------------------+
  | This file is a part of CiviCRM.                                    |
  |                                                                    |
@@ -28,9 +28,7 @@
 /**
  *
  * @package CRM
- * @copyright CiviCRM LLC (c) 2004-2015
- * $Id$
- *
+ * @copyright CiviCRM LLC (c) 2004-2017
  */
 class CRM_Contact_Page_DedupeRules extends CRM_Core_Page_Basic {
 
@@ -99,8 +97,6 @@ class CRM_Contact_Page_DedupeRules extends CRM_Core_Page_Basic {
    * This method is called after the page is created. It checks for the type
    * of action and executes that action. Finally it calls the parent's run
    * method.
-   *
-   * @return void
    */
   public function run() {
     // get the requested action, default to 'browse'
@@ -136,8 +132,6 @@ class CRM_Contact_Page_DedupeRules extends CRM_Core_Page_Basic {
 
   /**
    * Browse all rule groups.
-   *
-   * @return void
    */
   public function browse() {
     // get all rule groups
@@ -221,7 +215,11 @@ class CRM_Contact_Page_DedupeRules extends CRM_Core_Page_Basic {
 
     $rgDao = new CRM_Dedupe_DAO_RuleGroup();
     $rgDao->id = $id;
-    $rgDao->delete();
+    if ($rgDao->find(TRUE)) {
+      $rgDao->delete();
+      CRM_Core_Session::setStatus(ts("The rule '%1' has been deleted.", array(1 => $rgDao->title)), ts('Rule Deleted'), 'success');
+      CRM_Utils_System::redirect(CRM_Utils_System::url($this->userContext(), 'reset=1'));
+    }
   }
 
 }
