+++ profiles/civicrm_starterkit/modules/civicrm/CRM/Core/I18n/Schema.php
@@ -1,9 +1,9 @@
 <?php
 /*
  +--------------------------------------------------------------------+
- | CiviCRM version 4.6                                                |
+ | CiviCRM version 4.7                                                |
  +--------------------------------------------------------------------+
- | Copyright CiviCRM LLC (c) 2004-2015                                |
+ | Copyright CiviCRM LLC (c) 2004-2017                                |
  +--------------------------------------------------------------------+
  | This file is a part of CiviCRM.                                    |
  |                                                                    |
@@ -28,16 +28,12 @@
 /**
  *
  * @package CRM
- * @copyright CiviCRM LLC (c) 2004-2015
- * $Id$
- *
+ * @copyright CiviCRM LLC (c) 2004-2017
  */
 class CRM_Core_I18n_Schema {
 
   /**
    * Drop all views (for use by CRM_Core_DAO::dropAllTables() mostly).
-   *
-   * @return void
    */
   public static function dropAllViews() {
     $domain = new CRM_Core_DAO_Domain();
@@ -62,8 +58,6 @@ class CRM_Core_I18n_Schema {
    *
    * @param string $locale
    *   the first locale to create (migrate to).
-   *
-   * @return void
    */
   public static function makeMultilingual($locale) {
     $domain = new CRM_Core_DAO_Domain();
@@ -84,14 +78,18 @@ class CRM_Core_I18n_Schema {
       // drop old indices
       if (isset($indices[$table])) {
         foreach ($indices[$table] as $index) {
-          $queries[] = "DROP INDEX {$index['name']} ON {$table}";
+          if (CRM_Core_BAO_SchemaHandler::checkIfIndexExists($table, $index['name'])) {
+            $queries[] = "DROP INDEX {$index['name']} ON {$table}";
+          }
         }
       }
       // deal with columns
       foreach ($hash as $column => $type) {
         $queries[] = "ALTER TABLE {$table} ADD {$column}_{$locale} {$type}";
-        $queries[] = "UPDATE {$table} SET {$column}_{$locale} = {$column}";
-        $queries[] = "ALTER TABLE {$table} DROP {$column}";
+        if (CRM_Core_BAO_SchemaHandler::checkIfFieldExists($table, $column)) {
+          $queries[] = "UPDATE {$table} SET {$column}_{$locale} = {$column}";
+          $queries[] = "ALTER TABLE {$table} DROP {$column}";
+        }
       }
 
       // add view
@@ -117,8 +115,6 @@ class CRM_Core_I18n_Schema {
    *
    * @param string $retain
    *   the locale to retain.
-   *
-   * @return void
    */
   public static function makeSinglelingual($retain) {
     $domain = new CRM_Core_DAO_Domain();
@@ -163,8 +159,6 @@ class CRM_Core_I18n_Schema {
    *   schema structure class to use to recreate indices.
    *
    * @param array $triggers
-   *
-   * @return void
    */
   public static function makeSinglelingualTable(
     $retain,
@@ -244,8 +238,6 @@ class CRM_Core_I18n_Schema {
    *   the new locale to add.
    * @param string $source
    *   the locale to copy from.
-   *
-   * @return void
    */
   public static function addLocale($locale, $source) {
     // get the current supported locales
@@ -303,8 +295,6 @@ class CRM_Core_I18n_Schema {
    *   locales to be rebuilt.
    * @param string $version
    *   version of schema structure to use.
-   *
-   * @return void
    */
   public static function rebuildMultilingualSchema($locales, $version = NULL) {
     if ($version) {
@@ -375,7 +365,10 @@ class CRM_Core_I18n_Schema {
     global $dbLocale;
     $tables = self::schemaStructureTables();
     foreach ($tables as $table) {
-      $query = preg_replace("/([^'\"])({$table})([^_'\"])/", "\\1\\2{$dbLocale}\\3", $query);
+      // CRM-19093
+      // should match the civicrm table name such as: civicrm_event
+      // but must not match the table name if it's a substring of another table: civicrm_events_in_cart
+      $query = preg_replace("/([^'\"])({$table})(\z|[^a-z_'\"])/", "\\1\\2{$dbLocale}\\3", $query);
     }
     // uncomment the below to rewrite the civicrm_value_* queries
     // $query = preg_replace("/(civicrm_value_[a-z0-9_]+_\d+)([^_])/", "\\1{$dbLocale}\\2", $query);
